+++ src/gort/exceptions.py
@@ -9,10 +9,9 @@
 from __future__ import annotations
 
 import inspect
+from enum import Enum
 
 from typing import TYPE_CHECKING
-
-from gort.maskbits import ErrorCodes
 
 
 if TYPE_CHECKING:
@@ -23,6 +22,31 @@ if TYPE_CHECKING:
 
 def decapitalize_first_letter(s, upper_rest=False):
     return "".join([s[:1].lower(), (s[1:].upper() if upper_rest else s[1:])])
+
+
+class ErrorCodes(Enum):
+    """List of error codes."""
+
+    UNCATEGORISED_ERROR = 0
+    NOT_IMPLEMENTED = 1
+    COMMAND_FAILED = 2
+    COMMAND_TIMEDOUT = 3
+    USAGE_ERROR = 4
+    TIMEOUT = 5
+    CANNOT_MOVE_LOCAL_MODE = 101
+    FAILED_REACHING_COMMANDED_POSITION = 102
+    INVALID_TELESCOPE_POSITION = 103
+    FIBSEL_INVALID_POSITION = 201
+    SECTROGRAPH_FAILED_EXPOSING = 301
+    SECTROGRAPH_NOT_IDLE = 302
+    INVALID_CALIBRATION_SEQUENCE = 303
+    LOCAL_MODE_FAILED = 501
+    DOOR_STATUS_FAILED = 502
+    INVALID_PIXEL_NAME = 610
+    SCHEDULER_UNCATEGORISED = 701
+    SCHEDULER_CANNOT_FIND_TILE = 702
+    ACQUISITION_FAILED = 801
+    UNKNOWN_ERROR = 999
 
 
 class GortError(Exception):